save and reuse pull request message if creating it failed

+++ lib/hub/commands.rb
@@ -218,6 +218,8 @@ module Hub
 
       pull = api_client.create_pullrequest(options)
 
+      delete_pullreq_editmsg_file
+
       args.executable = 'echo'
       args.replace [pull['html_url']]
     rescue GitHubAPI::Exceptions
@@ -995,19 +997,24 @@ help
     end
 
     def pullrequest_editmsg(changes)
-      message_file = File.join(git_dir, 'PULLREQ_EDITMSG')
-      File.open(message_file, 'w') { |msg|
-        yield msg
-        if changes
-          msg.puts "#\n# Changes:\n#"
-          msg.puts changes.gsub(/^/, '# ').gsub(/ +$/, '')
-        end
-      }
+      message_file = pullreq_editmsg_file
+
+      unless File.exists?(message_file)
+        File.open(message_file, 'w') { |msg|
+          yield msg
+          if changes
+            msg.puts "#\n# Changes:\n#"
+            msg.puts changes.gsub(/^/, '# ').gsub(/ +$/, '')
+          end
+        }
+      end
+
       edit_cmd = Array(git_editor).dup
       edit_cmd << '-c' << 'set ft=gitcommit' if edit_cmd[0] =~ /^[mg]?vim$/
       edit_cmd << message_file
       system(*edit_cmd)
       abort "can't open text editor for pull request message" unless $?.success?
+
       title, body = read_editmsg(message_file)
       abort "Aborting due to empty pull request title" unless title
       [title, body]
@@ -1026,6 +1033,14 @@ help
       body.strip!
 
       [title =~ /\S/ ? title : nil, body =~ /\S/ ? body : nil]
+    end
+
+    def delete_pullreq_editmsg_file
+      File.delete(pullreq_editmsg_file) if File.exists?(pullreq_editmsg_file)
+    end
+
+    def pullreq_editmsg_file
+      File.join(git_dir, 'PULLREQ_EDITMSG')
     end
 
     def expand_alias(cmd)
